+++ src/roguelike/__init__.py
@@ -1,3 +1,7 @@
 from .title import (
     Title,
 )
+
+from .window import (
+    Window,
+)
