+++ tests/unit/test_title.py
@@ -1,6 +1,6 @@
 # coding: utf-8
 
-from roguelike.roguelike import MessageEngine
+from roguelike.main import MessageEngine
 from roguelike.title import Title
 
 def test_update():
